requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
distro==1.9.0
exceptiongroup==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.10.0
joblib==1.5.1
lxml==6.0.0
newspaper3k==0.2.8
nltk==3.9.1
openai==1.97.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-telegram-bot==22.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
schedule==1.2.2
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
